fix: fixed exec-on-workspace-changed bug by bypassing the callback
aerospace's exec-on-workspace-changed sometimes fires 3 times, ping
ponging for a single workspace change.

This mostly happens when my browser(Zen) is focused on a text input?
Although I am pretty sure this is a bug since I had no issues with v0.15

Anyways, this fixes the issue by not using the callback!