refactor(sketchybar): migrated from bash scripts to SBarLua
The motivation for this is so i can upgrade to the latest version of
aerospace(0.19.2), but this version has a bug where
aerospace_workspace_change fires multiple times causing the space
bashscript to mix up the sketchybar state dude to a racecondition.
[bug](https://github.com/nikitabobko/AeroSpace/discussions/1807)

The solution: redraw the whole bar on every event. But this is really
really slow in bash, so a complete overhaul to lua assisted with claude
was the solution!

most things are 1:1. Spaces obviously work very differently, with a
state diff system to reduce calls to sketchybar.
Only other thing that changed is memswap now only shows when swap is
actually being used.

- removed old bash files

.

- updated readme